Reorganise page expectation helpers into a single support file
This helps us to avoid duplication of the helper methods across different feature specs.
