+++ psd-web/spec/features/add_test_results_spec.rb
@@ -63,14 +63,6 @@ RSpec.feature "Adding a test result", :with_stubbed_elasticsearch, :with_stubbed
     end
   end
 
-  def expect_to_be_on_new_activity_page
-    expect_page_to_have_h1("New activity")
-  end
-
-  def expect_to_be_on_record_test_result_page
-    expect_page_to_have_h1("Record test result")
-  end
-
   def fill_in_test_result_submit_form(legislation:, date:, test_result:, file:)
     select legislation, from: "test_legislation"
     fill_in "Day",   with: date.day if date
