+++ psd-web/spec/features/create_enquiry_as_opss_user_spec.rb
@@ -97,28 +97,6 @@ RSpec.feature "Reporting enquiries", :with_stubbed_elasticsearch, :with_stubbed_
     end
   end
 
-  def expect_to_be_on_about_enquiry_page
-    expect(page).to have_current_path("/enquiry/about_enquiry")
-    expect_page_to_have_h1("New enquiry")
-  end
-
-  def expect_to_be_on_complainant_page
-    expect(page).to have_current_path("/enquiry/complainant")
-    expect_page_to_have_h1("New enquiry")
-  end
-
-  def expect_to_be_on_complainant_details_page
-    expect(page).to have_current_path("/enquiry/complainant_details")
-    expect_page_to_have_h1("New enquiry")
-    expect(page).to have_css(".govuk-fieldset__legend--m", text: "What are their contact details?")
-  end
-
-  def expect_to_be_on_enquiry_details_page
-    expect(page).to have_current_path("/enquiry/enquiry_details")
-    expect_page_to_have_h1("New enquiry")
-    expect(page).to have_css(".govuk-fieldset__legend--m", text: "What is the enquiry?")
-  end
-
   def expect_details_on_summary_page(contact_name:, contact_email:, contact_phone:)
     expect(page.find("dt", text: "Source type")).to have_sibling("dd", text: "Consumer")
     expect(page).to have_css("p", text: contact_name)
