+++ psd-web/spec/features/delete_attachment_from_case_spec.rb
@@ -19,7 +19,7 @@ RSpec.feature "Deleting an attachment from a case", :with_stubbed_elasticsearch,
 
     click_button "Delete attachment"
 
-    expect_to_be_on_case_overview_page
+    expect_to_be_on_case_page(case_id: investigation.pretty_id)
 
     click_link "Attachments"
 
@@ -28,24 +28,6 @@ RSpec.feature "Deleting an attachment from a case", :with_stubbed_elasticsearch,
     click_link "Activity"
 
     expect_case_activity_page_to_show_deleted_document
-  end
-
-  def expect_to_be_on_attachments_page
-    expect(page).to have_current_path("/cases/#{investigation.pretty_id}/attachments")
-    expect(page).to have_selector("h1", text: "Attachments")
-    expect(page).to have_selector("h2", text: document.title)
-    expect(page).to have_selector("p",  text: document.description)
-  end
-
-  def expect_to_be_on_remove_attachment_confirmation_page
-    expect(page).to have_current_path("/cases/#{investigation.pretty_id}/documents/#{document.id}/remove")
-    expect(page).to have_selector("h2", text: "Remove attachment")
-    expect(page).to have_link("Back", href: "/cases/#{investigation.pretty_id}")
-  end
-
-  def expect_to_be_on_case_overview_page
-    expect(page).to have_current_path("/cases/#{investigation.pretty_id}")
-    expect(page).to have_selector("h1", text: "Overview")
   end
 
   def expect_remove_attachment_confirmation_page_to_show_attachment_information
