+++ psd-web/spec/features/delete_attachment_from_product_spec.rb
@@ -18,7 +18,7 @@ RSpec.feature "Deleting an attachment from a product", :with_stubbed_elasticsear
 
     click_link "Remove document"
 
-    expect_to_be_on_remove_attachment_confirmation_page
+    expect_to_be_on_remove_attachment_from_product_confirmation_page
     expect_remove_attachment_confirmation_page_to_show_attachment_information
 
     click_button "Delete attachment"
@@ -28,23 +28,6 @@ RSpec.feature "Deleting an attachment from a product", :with_stubbed_elasticsear
     click_link "Attachments"
 
     expect_product_attachments_page_not_to_show_deleted_attachment
-  end
-
-  def expect_to_be_on_product_attachments_page
-    expect(page).to have_selector("h2", text: "Attachments")
-    expect(page).to have_selector("h2", text: document.title)
-    expect(page).to have_selector("p",  text: document.description)
-  end
-
-  def expect_to_be_on_remove_attachment_confirmation_page
-    expect(page).to have_current_path("/products/#{product.id}/documents/#{document.id}/remove")
-    expect(page).to have_selector("h2", text: "Remove attachment")
-    expect(page).to have_link("Back", href: "/products/#{product.id}")
-  end
-
-  def expect_to_be_on_product_page
-    expect(page).to have_current_path("/products/#{product.id}")
-    expect(page).to have_selector("h1", text: product.name)
   end
 
   def expect_remove_attachment_confirmation_page_to_show_attachment_information
