+++ psd-web/spec/features/edit_attachment_spec.rb
@@ -27,12 +27,6 @@ RSpec.feature "Editing an attachment on a case", :with_stubbed_elasticsearch, :w
     expect_case_activity_page_to_show_entered_information
   end
 
-  def expect_to_be_on_edit_attachment_page
-    expect(page).to have_current_path("/cases/#{investigation.pretty_id}/documents/#{document.to_param}/edit")
-    expect(page).to have_selector("h2", text: "Edit document details")
-    expect(page).to have_link("Back", href: "/cases/#{investigation.pretty_id}")
-  end
-
   def expect_case_attachments_page_to_show_entered_information
     expect(page).to have_selector("h2", text: new_title)
     expect(page).to have_selector("p",  text: new_description)
