+++ psd-web/spec/features/invite_user_spec.rb
@@ -27,8 +27,4 @@ RSpec.feature "Inviting a user", :with_stubbed_mailer, :with_stubbed_elasticsear
       expect(page).to have_css(".govuk-error-summary__list", text: "Email address belongs to a user that has been deleted. Email OPSS if you would like their account restored.")
     end
   end
-
-  def expect_to_be_on_invite_a_team_member_page
-    expect(page).to have_css("h1", text: "Invite a team member")
-  end
 end
