+++ psd-web/spec/support/feature_helpers.rb
@@ -13,11 +13,6 @@ def expect_page_to_have_h1(header)
   expect(page).to have_css("h1", text: header)
 end
 
-def expect_to_be_on_coronavirus_page(path)
-  expect(page).to have_current_path(path)
-  expect(page).to have_selector("h1", text: "Is this case related to the coronavirus outbreak?")
-end
-
 def enter_contact_details(contact_name:, contact_email:, contact_phone:)
   fill_in "complainant[name]", with: contact_name
   fill_in "complainant_email_address", with: contact_email
